Extract SVG injection setup and implementation to separate function
