+++ src/render/renderD3.js
@@ -1,5 +1,7 @@
 import * as d3 from 'd3';
 import Bezier from 'bezier-js';
+import SVGInjectInstance from '@iconfu/svg-inject';
+
 import bezierBuilder from '../utils/bezierBuilder';
 import { clearSVG } from '../utils/svgTools';
 
@@ -307,4 +309,69 @@ export const createUnknownText = (svg, index, element) => {
     })
     .attr('startOffset', '50%')
     .text(`Interest details unknown`);
+};
+
+export const setNodeLabelBkg = (color) => {
+  d3.selectAll('.edgeLabels .edgeLabel .label, .nodes .node .label').each(function (d, i) {
+    const label = d3.select(this);
+    const text = label.select('text');
+    const textParent = text.select(function () {
+      return this.parentNode;
+    });
+    const bBox = text.node().getBBox();
+
+    textParent
+      .insert('rect', ':first-child')
+      .attr('x', bBox.x)
+      .attr('y', bBox.y)
+      .attr('height', bBox.height)
+      .attr('width', bBox.width)
+      .style('fill', color);
+  });
+};
+
+export const injectSVGElements = (imagesPath, inner, g) => {
+  // create the nodetype injectable img elements used by an image injection library later
+  inner.selectAll('g.node').each(function (d, i) {
+    if (g.node(d).nodeType !== null) {
+      d3.select(this)
+        .append('img')
+        .attr('width', 120)
+        .attr('height', 120)
+        .attr('x', -60)
+        .attr('y', -60)
+        .attr('class', 'node-label label-container injectable')
+        .attr('src', function (d) {
+          return `${imagesPath}/${g.node(d).nodeType}`;
+        });
+    }
+  });
+
+  // create the flag injectable img elements
+  inner.selectAll('g.node').each(function (d, i) {
+    if (g.node(d).countryCode !== null) {
+      d3.select(this)
+        .append('img')
+        .attr('src', function (el) {
+          return `${imagesPath}/flags/${g.node(el).countryCode.toLowerCase()}.svg`;
+        })
+        .attr('width', '75')
+        .attr('height', '50')
+        .attr('x', '10')
+        .attr('y', '-80')
+        .attr('class', 'injectable flag');
+    }
+  });
+
+  if (document.querySelector('img.injectable')) {
+    // this allows SVGs to be injected directly from source
+    SVGInjectInstance(document.querySelectorAll('img.injectable')).then(() => {
+      d3.selectAll('.flag')
+        .insert('rect', ':first-child')
+        .attr('class', 'flag-bg')
+        .attr('height', '100%')
+        .attr('width', '100%')
+        .attr('style', 'fill: white; stroke: black; stroke-width: 2px;');
+    });
+  }
 };
